uncaught typo
no test for this according to coverage report, so unsurprising that it didn't get caught. I copy/pasted it for something else and ran into the `AttributeError`. 
I performed a `grep -rni` on that string and this seems to be the only place it happens.
```
[BET]$ grep -rni 'sset_.domain'
bet/.ipynb_checkpoints/sample-checkpoint.py:1753:        if self._domain is not None and sset_.domain is not None:
bet/sample.py.bak:1751:        if self._domain is not None and sset_.domain is not None:
bet/sample.py:1753:        if self._domain is not None and sset_.domain is not None:
build/lib/bet/sample.py:1753:        if self._domain is not None and sset_.domain is not None:
```